+++ package/mirage-core/src/Mirage/Core/Audio/Microphone/Detection.fs
@@ -69,7 +69,7 @@ type VoiceDetector<'State> =
 
 /// Initialize a vad detector by providing a vad algorithm, an action to
 /// perform when speech is detected, as well as a source to read samples from.
-let VoiceDetector<'State> minSilenceDurationMs (stopDetection: 'State -> bool) (detectSpeech: DetectVoice) (onVoiceDetected: DetectAction -> Async<Unit>) =
+let VoiceDetector<'State> minSilenceDurationMs (ignoreDetection: 'State -> bool) (detectSpeech: DetectVoice) (onVoiceDetected: DetectAction -> Async<Unit>) =
     let minSilenceSamples = float32 SamplingRate * float32 minSilenceDurationMs / 1000f
     let agent = new BlockingQueueAgent<ValueTuple<'State, ResampledAudio>>(Int32.MaxValue)
     let samples =
@@ -87,8 +87,8 @@ let VoiceDetector<'State> minSilenceDurationMs (stopDetection: 'State -> bool) (
             samples.original.AddRange currentAudio.original.samples
             samples.resampled.AddRange currentAudio.resampled.samples
             let! probability =
-                if stopDetection state then
-                    result 0f
+                if ignoreDetection state then
+                    result 1f
                 else
                     detectSpeech currentAudio.resampled.samples
             let fullAudio =