+++ package/mirage-lc/src/Mirage/Domain/Config.fs
@@ -145,7 +145,7 @@ type LocalConfig(general: ConfigFile, enemies: ConfigFile) =
     member val MinSilenceDurationMs =
         let description =
             "The minimum amount of detected silence before a voice clip should stop recording. By default, a recording stops after 2 seconds of non-detected speech is found.\n"
-                + "Note: This config option is for advanced users only. If you don't know what you're doing, don't touch this option."
+                + "Note: This config option is for advanced users only. If you don't know what you're doing, don't touch this option. This only affects voice activity users."
         bindAdvanced
             "Minimum silence duration (in milliseconds)"
             2000
@@ -154,7 +154,7 @@ type LocalConfig(general: ConfigFile, enemies: ConfigFile) =
     member val MinAudioDurationMs =
         let description =
             "The minimum amount of audio a voice clip should contain. If you have a lot of saved audio clips that contains useless audio, you can consider increasing this value.\n"
-                + "Note: This config option is for advanced users only. If you don't know what you're doing, don't touch this option."
+                + "Note: This config option is for advanced users only. If you don't know what you're doing, don't touch this option. This only affects voice activity users."
         bindAdvanced
             "Minimum audio duration (in milliseconds)"
             150